steps refactor
